requirements.txt
kafka-python==2.0.2
pandas==2.2.3
scikit-learn==1.5.2
